requirements-streamlit.txt
# Streamlit App Requirements
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0
